Add pressure plate enemy spawning.
Code is added and tested, but currently commented out as not tested in all scenes.
When pressure plate is stepped on, it disappears and enemies spawn in.
Doing this on laptop and dang it is hard to do with really slow loading speeds and small frame rate when playing.
Might move grounded check to fixed update as double jump doesn't work on here due to slow speed of updates on laptop.