Adjust tutorial coding.
Door 3 trigger is now at the top of the ramp rather than at the bottom.
The double jump is now deactivated when the player is in the dash section. Once the player dashes through the section, they get their double jump back.
Enemies now spawn 2 seconds after first door is open and every 5 seconds after that.