Fix double jumping when player shouldn't if jump was spammed quick enough.
The cooldown before the player could jump again was 0.25 seconds. With this, the bool that allowed the player to jump again could not become false quick enough, so if the player hit jump twice in 0.25 seconds, they jumped twice.
Value is now 0.3 seconds, and I can't jump twice. Maybe a goated button masher could. Also, do people even read this part of the commit? Let me know :D
When double jump is active, cool down is only 0.05 seconds so the double jump can occur quickly. Can't triple jump in this scenario, so I think it's good.