Add all door functionality to tutorial. Tutorial done I believe.
All of the doors in the tutorial now operate based on the conditions they need.
Also, player prefab now includes a trigger collider so that the player can interact with doors in other scenes. This issue was stumping me for a couple hours, but then I realized the collider on the player was only in the tutorial, not the other scenes.