Create player functionality with moving platforms. Adjust enemy placements.
Player now sticks to moving platforms when landing on them. This is done through setting the platform as the player's parent and adjusting the player's velocity to match that of the platform.
Adjust enemy placement by making all enemies the child of the platform they are standing on. Doing this ensures that the enemy adheres to the movement of the platform, just like the player.