Adjust tutorial PB logging.
New PBs and leaderboard entries only occur now if the player has cleared most of the tutorial. The condition to log entries is now if the player can double jump and dash. Double jump and dash means the player has unlocked all mechanics, so they will be allowed to log an entry.
This prevents logging entries when player just goes in the tutorial to access a portal rather than complete the tutorial.