Add death zones and adjust player respawn when falling off level.
When player falls, rather than checking for a y position, the player now respawns if they hit a death zone.
Death zones are empty objects with just a box collider and have the DeathZone tag. The tag is very very important.
If a player hits an object with the death zone tag, the player will respawn.
Death zones can be placed wherever by the design team and the functionality will still work as long as the death zone uses the DeathZone tag and has a collider with is trigger checked.