Adjust leaderboard UI button looks and function.
Leaderboard button is now plural because there are multiple.
The leaderboard buttons on the results screen now looks like it should be there.
When exiting leaderboard, it now goes to leaderboard selection screen, but goes to menu from leaderboard selection screen. This is because it is easier to access multiple leaderboards if you weren't sent to the menu each time.
Temp frame rate change keys have new binds. It is now hold minus (decrease) or equals (increase) and press F to decrease or increase by an interval of 5 frames.