Huge settings overhaul.
Settings menu is now its own prefab for easy implementation. The features within the settings menu are on self contained, so no buttons need to be assigned when adding the prefab to the scene.
Included in making settings menu self contained is creating a settings menu script. The main menu and game manager no longer have redundant code as all of that code is now within a settings menu script.
Every slider now has a value displayed next to it so players can get an idea of where there slider is or was.
Changed settings scroll view so the player can choose a category and then open a settings menu from there.
There is now a performance settings menu in which the player can apply VSync and set their FPS limit. The FPS display has been moved into here. Regarding the FPS limit, it is currently a slider, but should it be a drop down list?
Controls menu now exists, but is very incomplete. In the future, it will have a screen of controls that the player can change if they desire.
Accidently deleted the crosshair setting, so will be adding that back in the future as well.